Rename init command to initialize
