+++ pkg/app/pipectl/cmd/initialize/initialize.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package init
+package initialize
 
 import (
 	"context"
